Перевести проект на архитектурный паттерн MVVM.Задание 33.7
