--- app/src/main/java/com/example/myapplication/HomeFragment.kt
@@ -1,171 +1,0 @@
-package com.example.myapplication
-
-import android.content.Intent
-import android.os.Bundle
-import android.transition.Scene
-import android.transition.Slide
-import android.transition.TransitionManager
-import android.transition.TransitionSet
-import android.view.Gravity
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.appcompat.widget.SearchView
-import androidx.constraintlayout.motion.widget.MotionLayout
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import com.example.myapplication.databinding.ActivityMainBinding
-import com.example.myapplication.databinding.FragmentHomeBinding
-import com.example.myapplication.databinding.HomeFragmentMotionSceneBinding
-import java.util.Locale
-
-class HomeFragment : Fragment() {
-    //private lateinit var binding: MergeHomeScreenContentBinding
-    private lateinit var binding: HomeFragmentMotionSceneBinding
-    private lateinit var filmsAdapter: FilmListRecyclerAdapter
-
-    val filmsDataBase = listOf(
-        Film(
-            "Капитан Америка",
-            R.drawable.film_poster1,
-            "Сюжет разворачивается после фильма «Вечные» и сериала «Сокол и Зимний Солдат». Группа наёмников похищает образец металла Адамантий, добытого из тела Целестиала Тиамута. Но их останавливают новый Капитан Америка (Энтони Маки) и его помощник Сокол. Сразу после этих событий и неудачного покушения на нового президента США, коим стал бывший генерал Таддеус «Громовержец» Росс, начинают происходить загадочные события, коим из-за кулис руководит неизвестный кукловод. "
-        ,4.5f),
-        Film(
-            "Побег из Шоушенка",
-            R.drawable.film_poster2,
-            "Сюжет: успешный банкир Энди Дюфрейн оказывается главным обвиняемым в двойном убийстве. Одним вечером полиция обнаруживает тела его жены с любовником, а сам Энди просто не может вспомнить, что делал в ночь трагедии. Так он оказывается в самой суровой тюрьме Новой Англии, где ему предстоит провести оставшуюся жизнь.Вежливому и кроткому Энди непросто приспособиться к жестокой тюремной реальности, где опасность исходит не только от заключённых, но и от администрации. Но со временем Энди удаётся завести друзей, а его профессиональные качества талантливого банкира очень пригождаются, когда он начинает оказывать сотрудникам Шоушенка финансовые и юридические консультации."
-        , 7.7f),
-        Film(
-            "Форрест Гамп",
-            R.drawable.film_poster3,
-            " Сидя на автобусной остановке, Форрест Гамп — не очень умный, но добрый и открытый парень — рассказывает случайным встречным историю своей необыкновенной жизни. С самого малолетства парень страдал от заболевания ног, соседские мальчишки дразнили его, но в один прекрасный день Форрест открыл в себе невероятные способности к бегу. Подруга детства Дженни всегда его поддерживала и защищала, но вскоре дороги их разошлись."
-        ,9.9f),
-        Film(
-            "Бойцовский клуб",
-            R.drawable.film_poster4,
-            "Сотрудник страховой компании страдает хронической бессонницей и отчаянно пытается вырваться из мучительно скучной жизни. Однажды в очередной командировке он встречает некоего Тайлера Дёрдена — харизматического торговца мылом с извращенной философией. Тайлер уверен, что самосовершенствование — удел слабых, а единственное, ради чего стоит жить, — саморазрушение. Проходит немного времени, и вот уже новые друзья лупят друг друга почем зря на стоянке перед баром, и очищающий мордобой доставляет им высшее блаженство. Приобщая других мужчин к простым радостям физической жестокости, они основывают тайный Бойцовский клуб, который начинает пользоваться невероятной популярностью."
-        ,5.4f),
-        Film(
-            "Властелин колец: Две крепости",
-            R.drawable.film_poster5,
-            "Братство распалось, но Кольцо Всевластья должно быть уничтожено. Фродо и Сэм вынуждены довериться Голлуму, который взялся провести их к вратам Мордора. Громадная армия Сарумана приближается: члены братства и их союзники готовы принять бой. Битва за Средиземье продолжается."
-        ,6.1f),
-        Film(
-            "Матрица",
-            R.drawable.film_poster6,
-            "Жизнь Томаса Андерсона разделена на две части: днём он — самый обычный офисный работник, получающий нагоняи от начальства, а ночью превращается в хакера по имени Нео, и нет места в сети, куда он бы не смог проникнуть. Но однажды всё меняется. Томас узнаёт ужасающую правду о реальности."
-        ,7.4f),
-        Film(
-            "Интерстеллар",
-            R.drawable.film_poster7,
-            "Когда засуха, пыльные бури и вымирание растений приводят человечество к продовольственному кризису, коллектив исследователей и учёных отправляется сквозь червоточину (которая предположительно соединяет области пространства-времени через большое расстояние) в путешествие, чтобы превзойти прежние ограничения для космических путешествий человека и найти планету с подходящими для человечества условиями."
-        ,9.9f),
-        Film(
-            "Унесённые призраками",
-            R.drawable.film_poster8,
-            "Тихиро с мамой и папой переезжает в новый дом. Заблудившись по дороге, они оказываются в странном пустынном городе, где их ждет великолепный пир. Родители с жадностью набрасываются на еду и к ужасу девочки превращаются в свиней, став пленниками злой колдуньи Юбабы. Теперь, оказавшись одна среди волшебных существ и загадочных видений, Тихиро должна придумать, как избавить своих родителей от чар коварной старухи."
-        ,2.3f),
-        Film(
-            "1+1",
-            R.drawable.film_poster9,
-            "Сюжет: в результате несчастного случая богатый бизнесмен Филипп становится инвалидом. Помогая ему адаптироваться, врачи направляют к нему работников по уходу за больными. Из всех кандидатов на роль медбрата привередливый богач выбирает самого странного: молодого африканского эмигранта с судимостью. Непосредственный араб становится для скучающего Филиппа и ангелом-хранителем, и источником ежедневных приключений. Отношения двух мужчин из разных социальных слоёв перерастают в крепкую дружбу"
-        ,4.6f)
-    )
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = HomeFragmentMotionSceneBinding.inflate(layoutInflater)
-        // сама навигация
-        initHomeFragment()
-        return binding.root
-    }
-
-    private fun initHomeFragment() {
-        binding.mainRecycler.layoutManager = LinearLayoutManager(requireContext())
-        filmsAdapter =
-            FilmListRecyclerAdapter(object : FilmListRecyclerAdapter.OnItemClickListener {
-                override fun click(film: Film) {
-                    (requireActivity() as MainActivity).launchDetailsFragment(film)
-                }
-            })
-        filmsAdapter.addItems(filmsDataBase)
-        binding.mainRecycler.adapter = filmsAdapter
-        val decorator = TopSpacingItemDecoration(8)
-        binding.mainRecycler.addItemDecoration(decorator)
-
-        binding.searchView.setOnClickListener {
-            binding.searchView.isIconified = false
-        }
-        binding.searchView.setOnQueryTextListener(object : SearchView.OnQueryTextListener {
-            override fun onQueryTextSubmit(query: String?): Boolean {
-                return true
-            }
-
-            override fun onQueryTextChange(newText: String?): Boolean {
-                if (newText.isNullOrEmpty()) {
-                    filmsAdapter.addItems(filmsDataBase)
-                    return true
-                }
-                //Фильтруем список на поискк подходящих сочетаний
-                val result = filmsDataBase.filter {
-                    //Чтобы все работало правильно, нужно и запрос, и имя фильма приводить к нижнему регистру
-                    it.title.lowercase(Locale.getDefault())
-                        .contains(newText.lowercase(Locale.getDefault()))
-                }
-                //Добавляем в адаптер
-                filmsAdapter.addItems(result)
-                return true
-            }
-        })
-    }
-//добавляем анимацию
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        val motionLayout = binding.root
-        motionLayout.setTransitionListener(object: MotionLayout.TransitionListener{
-            override fun onTransitionStarted(
-                motionLayout: MotionLayout?,
-                startId: Int,
-                endId: Int
-            )
-            {}
-
-            override fun onTransitionChange(
-                motionLayout: MotionLayout?,
-                startId: Int,
-                endId: Int,
-                progress: Float
-            ) {
-                binding.searchView.translationY = 0F
-                binding.searchView.alpha = 1F
-            }
-
-            override fun onTransitionCompleted(motionLayout: MotionLayout?, currentId: Int) {
-
-            }
-
-            override fun onTransitionTrigger(
-                motionLayout: MotionLayout?,
-                triggerId: Int,
-                positive: Boolean,
-                progress: Float
-            ) {}
-        })//фиксация чтобы при скролле не съехала
-        binding.mainRecycler.addOnScrollListener(object : RecyclerView.OnScrollListener() {
-            override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
-                binding.searchView.translationY = 0f
-                binding.searchView.alpha = 1f
-            }
-        })//включаем на стартовый экран при первом запуске
-        motionLayout.post {
-            motionLayout.transitionToEnd()
-            binding.mainRecycler.isNestedScrollingEnabled = false
-        }
-
-    }
-}
-
