+++ app/src/main/java/com/example/myapplication/domain/Film.kt
@@ -1,6 +1,5 @@
-package com.example.myapplication
+package com.example.myapplication.domain
 
-import android.annotation.SuppressLint
 import android.os.Parcel
 import android.os.Parcelable
