+++ app/src/main/java/com/example/myapplication/utils/AnimationHelper.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.example.myapplication.utils
 
 import android.app.Activity
 import android.view.View
