+++ app/src/main/java/com/example/myapplication/view/customviews/RatingDonutView.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.example.myapplication.view.customviews
 
 import android.content.Context
 import android.graphics.Canvas
@@ -8,6 +8,7 @@ import android.graphics.RectF
 import android.graphics.Typeface
 import android.util.AttributeSet
 import android.view.View
+import com.example.myapplication.R
 
 class RatingDonutView @JvmOverloads constructor(context: Context, attributeSet: AttributeSet? = null) : View(context, attributeSet) {
     //Овал для рисования сегментов прогресс бара
