+++ app/src/main/java/com/example/myapplication/view/fragments/DetailsFragment.kt
@@ -1,15 +1,15 @@
-package com.example.myapplication
+package com.example.myapplication.view.fragments
 
 import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
+import com.example.myapplication.R
 //import com.example.myapplication.databinding.ActivityDetailsBinding
 import com.example.myapplication.databinding.FragmentDetailsBinding
-import com.example.myapplication.databinding.FragmentHomeBinding
+import com.example.myapplication.domain.Film
 
 class DetailsFragment : Fragment() {
         private lateinit var binding: FragmentDetailsBinding
