+++ app/src/main/java/com/example/myapplication/view/fragments/WatchLaterFragment.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.example.myapplication.view.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,6 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.myapplication.databinding.FragmentWatchLaterBinding
+import com.example.myapplication.utils.AnimationHelper
 
 class WatchLaterFragment : Fragment() {
     private var binding: FragmentWatchLaterBinding? = null
