+++ app/src/main/java/com/example/myapplication/view/rv_adapters/TopSpacingItemDecoration.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.example.myapplication.view.rv_adapters
 
 import android.content.res.Resources
 import android.graphics.Rect
