+++ app/src/main/res/layout/film_item.xml
@@ -18,7 +18,7 @@
             app:layout_constraintTop_toTopOf="parent"
             tools:src="@tools:sample/avatars" />
 
-    <com.example.myapplication.RatingDonutView
+    <com.example.myapplication.view.customviews.RatingDonutView
         android:id="@+id/rating_donut"
         android:layout_width="50dp"
         android:layout_height="50dp"
