+++ app/src/main/res/layout/fragment_home.xml
@@ -7,7 +7,7 @@
     tools:background="@drawable/background_search_view"
     android:id="@+id/home_fragment_root"
     android:visibility="visible"
-    tools:context=".HomeFragment"/>
+    tools:context=".view.fragments.HomeFragment"/>
